mocked a transparent note
